Refactor YParametersSettings into subcomponents
