+++ components/charts/EditModal/parameters/YParametersSettings.tsx
@@ -12,6 +12,7 @@ import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/comp
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
 import { Checkbox } from "@/components/ui/checkbox"
 import { X, Plus, ChevronDown, ChevronRight, Copy, Edit2 } from "lucide-react"
+import { YAxisGroup } from "./YAxisGroup"
 import { ChartComponent, InterlockDefinition } from "@/types"
 import { mockInterlockMaster } from "@/data/interlockMaster"
 import { InterlockRegistrationDialog } from "./InterlockRegistrationDialog"
@@ -347,354 +348,42 @@ export function YParametersSettings({ editingChart, setEditingChart }: YParamete
 
         <div className="flex-1 overflow-y-auto min-h-0">
           <div className="space-y-4">
-            {editingChart.yAxisParams && editingChart.yAxisParams.length > 0 ? (
-              Object.entries(groupParametersByAxis()).map(([axisNoStr, paramIndexes]) => {
-                const axisNo = parseInt(axisNoStr)
-                const firstParam = editingChart.yAxisParams![paramIndexes[0]]
-                const axisLabel = editingChart.yAxisLabels?.[axisNo] || ""
-                const axisRange = firstParam.range || { auto: true, min: 0, max: 100 }
-                
-                return (
-                  <div key={axisNo} className="border rounded-lg p-2 bg-muted/10">
-                    {/* Compact header: Y-Axis | Y-axis Label | Range | Axis No */}
-                    <div className="mb-1 pb-1 border-b">
-                      <div className="flex items-center gap-4">
-                        <div className="w-24">
-                          <h5 className="font-medium text-sm">Y-Axis {axisNo}</h5>
-                          <span className="text-xs text-muted-foreground">
-                            ({paramIndexes.length} param{paramIndexes.length !== 1 ? 's' : ''})
-                          </span>
-                        </div>
-                        
-                        <div className="flex-1">
-                          <Input
-                            value={axisLabel}
-                            onChange={(e) => updateAxisLabel(axisNo, e.target.value)}
-                            placeholder={`Y-axis ${axisNo} label`}
-                            className="h-7 text-sm"
-                          />
-                        </div>
-                        
-                        <Popover>
-                          <PopoverTrigger asChild>
-                            <Button variant="outline" size="sm" className="h-7 text-xs px-2">
-                              {axisRange.auto ? "Range: Auto" : `Range: ${axisRange.min} - ${axisRange.max}`}
-                            </Button>
-                          </PopoverTrigger>
-                          <PopoverContent className="w-80">
-                            <div className="space-y-3">
-                              <div className="flex items-center space-x-2">
-                                <Checkbox
-                                  id={`y-auto-axis-${axisNo}`}
-                                  checked={axisRange.auto}
-                                  onCheckedChange={(checked) => updateAxisRange(axisNo, { auto: checked === true })}
-                                />
-                                <Label htmlFor={`y-auto-axis-${axisNo}`} className="text-sm">Auto Range</Label>
-                              </div>
-                              <div className="space-y-2">
-                                <div>
-                                  <Label className="text-xs">Min Value</Label>
-                                  <Input
-                                    type="number"
-                                    value={axisRange.min}
-                                    onChange={(e) => updateAxisRange(axisNo, { min: parseFloat(e.target.value) || 0 })}
-                                    disabled={axisRange.auto}
-                                    className="h-8"
-                                  />
-                                </div>
-                                <div>
-                                  <Label className="text-xs">Max Value</Label>
-                                  <Input
-                                    type="number"
-                                    value={axisRange.max}
-                                    onChange={(e) => updateAxisRange(axisNo, { max: parseFloat(e.target.value) || 100 })}
-                                    disabled={axisRange.auto}
-                                    className="h-8"
-                                  />
-                                </div>
-                              </div>
-                            </div>
-                          </PopoverContent>
-                        </Popover>
-                        
-                        <div className="flex items-center gap-1">
-                          <Label className="text-xs">Axis No:</Label>
-                          <Input
-                            type="number"
-                            min="1"
-                            max="10"
-                            value={axisNo}
-                            onChange={(e) => {
-                              const newAxisNo = parseInt(e.target.value) || 1
-                              if (newAxisNo !== axisNo) {
-                                const newParams = [...(editingChart.yAxisParams || [])]
-                                paramIndexes.forEach(index => {
-                                  newParams[index] = { ...newParams[index], axisNo: newAxisNo }
-                                })
-                                setEditingChart({ ...editingChart, yAxisParams: newParams })
-                              }
-                            }}
-                            className="w-12 h-7 text-xs px-1"
-                          />
-                        </div>
-                      </div>
-                    </div>
-                    
-                    {/* Type | Parameter table */}
-                    <div className="space-y-0.5 pt-1">
-                      <div className="flex gap-2 mb-1 px-1 text-xs font-medium text-muted-foreground border-b pb-1">
-                        <div className="w-24">Type</div>
-                        <div className="flex-1">Parameter</div>
-                        <div className="w-7 flex justify-center">
-                          <Button
-                            variant="outline"
-                            size="sm"
-                            className="h-6 w-6 p-0 border-dashed border-blue-400 hover:border-blue-600 hover:bg-blue-50 text-blue-600 hover:text-blue-700"
-                            onClick={() => addParameterToAxis(axisNo)}
-                          >
-                            <Plus className="h-3.5 w-3.5" />
-                          </Button>
-                        </div>
-                      </div>
-                      
-                      {paramIndexes.map((index) => {
-                        const param = editingChart.yAxisParams![index]
-                        return (
-                          <div key={index} className="flex gap-2 items-center">
-                            <div className="w-24">
-                              <select
-                                ref={(el) => {
-                                  parameterTypeSelectRefs.current[index] = el
-                                }}
-                                value={param.parameterType || "Parameter"}
-                                onChange={(e) => handleParameterTypeChange(index, e.target.value as "Parameter" | "Formula" | "Interlock")}
-                                className="w-full h-7 px-2 py-1 border rounded-md text-xs"
-                              >
-                                <option value="Parameter">Parameter</option>
-                                <option value="Formula">Formula</option>
-                                <option value="Interlock">Interlock</option>
-                              </select>
-                            </div>
-                            
-                            <div className="flex-1">
-                              <div className="space-y-1">
-                                {param.parameterType === "Interlock" ? (
-                                  <TooltipProvider>
-                                    <Tooltip>
-                                      <TooltipTrigger asChild>
-                                        <div className="w-full">
-                                          <Popover 
-                                            open={openComboboxIndex === index} 
-                                            onOpenChange={(open) => {
-                                              setOpenComboboxIndex(open ? index : null)
-                                              setSearchQuery("")
-                                            }}
-                                          >
-                                            <PopoverTrigger asChild>
-                                              <Button
-                                                variant="outline"
-                                                role="combobox"
-                                                className="h-7 w-full justify-start text-sm font-normal min-w-0"
-                                              >
-                                                <span className="truncate text-left mr-auto">
-                                                  {param.parameter || "Select Interlock"}
-                                                </span>
-                                                <ChevronDown className="h-4 w-4 opacity-50 flex-shrink-0 ml-2" />
-                                              </Button>
-                                            </PopoverTrigger>
-                                            <PopoverContent className="w-[300px] p-0">
-                                              <Command shouldFilter={false}>
-                                                <CommandInput 
-                                                  placeholder="Search by name, plant, machine..." 
-                                                  value={searchQuery}
-                                                  onValueChange={setSearchQuery}
-                                                />
-                                                <CommandEmpty>No interlock found.</CommandEmpty>
-                                                <CommandGroup>
-                                                  {filterInterlocks(mockInterlockMaster).map((master) => (
-                                                    <CommandItem
-                                                      key={master.id}
-                                                      value={master.id}
-                                                      onSelect={() => handleInterlockSelect(index, master.id, "select")}
-                                                      className="flex items-center justify-between group"
-                                                    >
-                                                      <div className="flex flex-col items-start flex-1">
-                                                        <span className="font-medium text-left">{master.name}</span>
-                                                        <span className="text-xs text-muted-foreground text-left">
-                                                          {master.plant_name} • {master.machine_no}
-                                                        </span>
-                                                      </div>
-                                                      <div className="flex gap-1">
-                                                        <Button
-                                                          variant="ghost"
-                                                          size="sm"
-                                                          className="h-6 w-6 p-0 opacity-0 group-hover:opacity-100 transition-opacity"
-                                                          onClick={(e) => {
-                                                            e.stopPropagation()
-                                                            handleInterlockSelect(index, master.id, "edit")
-                                                          }}
-                                                          title="Edit interlock"
-                                                        >
-                                                          <Edit2 className="h-3 w-3" />
-                                                        </Button>
-                                                        <Button
-                                                          variant="ghost"
-                                                          size="sm"
-                                                          className="h-6 w-6 p-0 opacity-0 group-hover:opacity-100 transition-opacity"
-                                                          onClick={(e) => {
-                                                            e.stopPropagation()
-                                                            handleInterlockSelect(index, master.id, "duplicate")
-                                                          }}
-                                                          title="Duplicate interlock"
-                                                        >
-                                                          <Copy className="h-3 w-3" />
-                                                        </Button>
-                                                      </div>
-                                                    </CommandItem>
-                                                  ))}
-                                                  <CommandItem
-                                                    value="add-new"
-                                                    onSelect={() => handleInterlockSelect(index, "add-new")}
-                                                    className="border-t"
-                                                  >
-                                                    <Plus className="mr-2 h-4 w-4" />
-                                                    <span>Add new interlock...</span>
-                                                  </CommandItem>
-                                                </CommandGroup>
-                                              </Command>
-                                            </PopoverContent>
-                                          </Popover>
-                                        </div>
-                                      </TooltipTrigger>
-                                      <TooltipContent>
-                                        <p className="max-w-xs">
-                                          {param.parameter || "Select Interlock"}
-                                          {param.selectedThresholds && param.selectedThresholds.length > 0 && (
-                                            <>
-                                              <br />
-                                              <span className="text-xs">
-                                                Thresholds: {param.selectedThresholds.map(thresholdId => {
-                                                  const threshold = param.interlockDefinition?.thresholds.find(t => t.id === thresholdId)
-                                                  return threshold?.name
-                                                }).filter(Boolean).join(", ")}
-                                              </span>
-                                            </>
-                                          )}
-                                        </p>
-                                      </TooltipContent>
-                                    </Tooltip>
-                                  </TooltipProvider>
-                                ) : (
-                                  <Input
-                                    ref={(el) => {
-                                      parameterInputRefs.current[index] = el
-                                    }}
-                                    value={param.parameter}
-                                    onChange={(e) => {
-                                      const newParams = [...(editingChart.yAxisParams || [])]
-                                      newParams[index] = { ...newParams[index], parameter: e.target.value }
-                                      setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                    }}
-                                    placeholder="Parameter"
-                                    className="h-7 text-sm"
-                                  />
-                                )}
-                                
-                                {param.parameterType === "Interlock" && ((param.selectedThresholds && param.selectedThresholds.length > 0) || param.interlockDefinition) && (
-                                  <div className="flex gap-1 flex-wrap items-center mt-1">
-                                    {param.selectedThresholds?.map(thresholdId => {
-                                      const threshold = param.interlockDefinition?.thresholds.find(t => t.id === thresholdId)
-                                      return threshold ? (
-                                        <Badge 
-                                          key={thresholdId} 
-                                          variant="secondary" 
-                                          className="text-xs px-1.5 py-0 h-5 flex items-center gap-1"
-                                          style={{ 
-                                            backgroundColor: threshold.color + '20',
-                                            borderColor: threshold.color,
-                                            color: threshold.color
-                                          }}
-                                        >
-                                          <span>{threshold.name}</span>
-                                          <button
-                                            onClick={(e) => {
-                                              e.stopPropagation()
-                                              handleThresholdRemove(index, thresholdId)
-                                            }}
-                                            className="hover:bg-black/10 rounded-full p-0.5"
-                                          >
-                                            <X className="h-2 w-2" />
-                                          </button>
-                                        </Badge>
-                                      ) : null
-                                    })}
-                                    {param.interlockDefinition && (
-                                      <Popover>
-                                        <PopoverTrigger asChild>
-                                          <Button
-                                            variant="ghost"
-                                            size="sm"
-                                            className="h-5 w-5 p-0 border border-dashed border-gray-400 hover:border-gray-600"
-                                          >
-                                            <Plus className="h-3 w-3" />
-                                          </Button>
-                                        </PopoverTrigger>
-                                        <PopoverContent className="w-48 p-1">
-                                          <div className="space-y-1">
-                                            {param.interlockDefinition.thresholds
-                                              .filter(threshold => !param.selectedThresholds?.includes(threshold.id))
-                                              .map(threshold => (
-                                                <button
-                                                  key={threshold.id}
-                                                  onClick={() => handleThresholdAdd(index, threshold.id)}
-                                                  className="w-full text-left px-2 py-1 text-xs hover:bg-muted rounded flex items-center gap-2"
-                                                >
-                                                  <span
-                                                    className="w-2 h-2 rounded-full"
-                                                    style={{ backgroundColor: threshold.color }}
-                                                  />
-                                                  {threshold.name}
-                                                </button>
-                                              ))}
-                                            {param.interlockDefinition.thresholds.every(threshold => 
-                                              param.selectedThresholds?.includes(threshold.id)
-                                            ) && (
-                                              <div className="px-2 py-1 text-xs text-muted-foreground">
-                                                All thresholds selected
-                                              </div>
-                                            )}
-                                          </div>
-                                        </PopoverContent>
-                                      </Popover>
-                                    )}
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                            
-                            <div className="w-7">
-                              <Button
-                                variant="ghost"
-                                size="sm"
-                                className="h-7 w-7 p-0"
-                                onClick={() => {
-                                  const newParams = editingChart.yAxisParams?.filter((_, i) => i !== index) || []
-                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                }}
-                              >
-                                <X className="h-3 w-3" />
-                              </Button>
-                            </div>
-                          </div>
-                        )
-                      })}
-                    </div>
-                  </div>
-                )
-              })
-            ) : (
-              <p className="text-sm text-muted-foreground px-1">No Y parameters added yet.</p>
-            )}
+              {editingChart.yAxisParams && editingChart.yAxisParams.length > 0 ? (
+                Object.entries(groupParametersByAxis()).map(([axisNoStr, paramIndexes]) => {
+                  const axisNo = parseInt(axisNoStr)
+                  const firstParam = editingChart.yAxisParams![paramIndexes[0]]
+                  const axisLabel = editingChart.yAxisLabels?.[axisNo] || ""
+                  const axisRange = firstParam.range || { auto: true, min: 0, max: 100 }
+
+                  return (
+                    <YAxisGroup
+                      key={axisNo}
+                      axisNo={axisNo}
+                      paramIndexes={paramIndexes}
+                      axisLabel={axisLabel}
+                      axisRange={axisRange}
+                      editingChart={editingChart}
+                      setEditingChart={setEditingChart}
+                      updateAxisLabel={updateAxisLabel}
+                      updateAxisRange={updateAxisRange}
+                      addParameterToAxis={addParameterToAxis}
+                      parameterInputRefs={parameterInputRefs}
+                      parameterTypeSelectRefs={parameterTypeSelectRefs}
+                      openComboboxIndex={openComboboxIndex}
+                      setOpenComboboxIndex={setOpenComboboxIndex}
+                      searchQuery={searchQuery}
+                      setSearchQuery={setSearchQuery}
+                      handleParameterTypeChange={handleParameterTypeChange}
+                      handleInterlockSelect={handleInterlockSelect}
+                      filterInterlocks={filterInterlocks}
+                      handleThresholdRemove={handleThresholdRemove}
+                      handleThresholdAdd={handleThresholdAdd}
+                    />
+                  )
+                })
+              ) : (
+                <p className="text-sm text-muted-foreground px-1">No Y parameters added yet.</p>
+              )}
           </div>
         </div>
             </div>
