+++ components/charts/EditModal/parameters/index.ts
@@ -8,3 +8,5 @@ export { InterlockFormFields } from './InterlockFormFields'
 export { ThresholdColorSection } from './ThresholdColorSection'
 export { ThresholdPointsTable } from './ThresholdPointsTable'
 export { InterlockSection } from './InterlockSection'
+export { ParameterRow } from './ParameterRow'
+export { YAxisGroup } from './YAxisGroup'
